Refactor NotaRemisionModel and SocioNegocioModel to use double for monetary values; implement getNotasRemision in DepositoChequesImpl; enhance DepositoChequeRegisterScreen with responsive design and image upload functionality; add ClienteSearchDialog for client selection; update pubspec.yaml and pubspec.lock for new dependencies.

+++ lib/data/models/nota_remision_model.dart
@@ -18,8 +18,8 @@ class NotaRemisionModel {
   final int docNum;
   final DateTime fecha;
   final int numFact;
-  final int totalMonto;
-  final int saldoPendiente;
+  final double totalMonto;
+  final double saldoPendiente;
   final int audUsuario;
   final String codCliente;
   final String nombreCliente;
