+++ lib/domain/entities/nota_remision_entity.dart
@@ -4,8 +4,8 @@ class NotaRemisionEntity {
     final int docNum;
     final DateTime fecha;
     final int numFact;
-    final int totalMonto;
-    final int saldoPendiente;
+    final double totalMonto;
+    final double saldoPendiente;
     final int audUsuario;
     final String codCliente;
     final String nombreCliente;
